Fix delete in type reservation

+++ Reservation-app/src/main/java/ma/youcode/reservation/DAO/TypeReservationRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TypeReservationRepository extends JpaRepository<TypereservationEntity, Long> {
+public interface TypeReservationRepository extends JpaRepository<TypereservationEntity, Integer> {
 }
